Move buffer processing out of GstPipeline and NvDsPipeline
